+++ tests/testthat/test-Utility.R
@@ -183,7 +183,7 @@ test_that("Same Ratio test", {
   a_n <- c(-0.1, 0.1, 0.1, 0.2)
   control <- list("ncores" = 1, "maxiter" = 10, "verbose" = 0)
   alternative_model <- CoxRun(Cox(a, b, c) ~ plinear(d * d, 0) + loglinear(factor(e)), df, control = control, a_n = a_n, norm = "max", keep_constant = c(0, 1, 0))
-  expect_identical(Likelihood_Ratio_Test(alternative_model, alternative_model)$Difference, 0)
+  expect_error(Likelihood_Ratio_Test(alternative_model, alternative_model))
 })
 test_that("No Data Ratio test", {
   a <- list("baditem" = -300)